# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is a leards language learning app api.",
        "title": "Leards Backend API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/folders/{folder_id}/decks": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "creates a new deck in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decks"
                ],
                "summary": "Create a new deck",
                "operationId": "createNewDeck",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folder_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Create deck data",
                        "name": "createDeckData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/CreateDeckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DeckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/folders/{folder_id}/decks/{deck_id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "fetches the deck from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decks"
                ],
                "summary": "Get the deck by id",
                "operationId": "getDeckById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folder_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Deck ID",
                        "name": "deck_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DeckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "updates the deck in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decks"
                ],
                "summary": "Updates the deck by id",
                "operationId": "updateDeckById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folder_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Deck ID",
                        "name": "deck_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update deck data",
                        "name": "updateDeckData",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/UpdateDeckRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/DeckResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "deletes the deck in the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "decks"
                ],
                "summary": "Delete the deck by id",
                "operationId": "deleteDeckById",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folder_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Deck ID",
                        "name": "deck_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        },
        "/folders/{folder_id}/decks/{deck_id}/cards": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "fetches cards of the deck from the database",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Get all deck's cards",
                "operationId": "getCardsByDeckId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folder_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Deck ID",
                        "name": "deck_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/CardsResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "adds card without id, updates card with id, deletes card if it's not presented inside the request",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "cards"
                ],
                "summary": "Synchronizes cards",
                "operationId": "syncCardsByDeckId",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Folder ID",
                        "name": "folder_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "Deck ID",
                        "name": "deck_id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Sync cards data",
                        "name": "syncCardsRequest",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/SyncCardsRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/BasicResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/HTTPError"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "BasicResponse": {
            "type": "object",
            "properties": {
                "message": {
                    "type": "string",
                    "example": "Successfully"
                }
            }
        },
        "Card": {
            "type": "object",
            "properties": {
                "backSide": {
                    "type": "string"
                },
                "cardId": {
                    "type": "string"
                },
                "frontSide": {
                    "type": "string"
                }
            }
        },
        "CardsResponse": {
            "type": "object",
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Card"
                    }
                },
                "message": {
                    "type": "string",
                    "example": "Successfully"
                }
            }
        },
...
        "CreateDeckRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "name": {
                    "type": "string",
                    "example": "My new folder"
                },
                "parentFolderId": {
                    "type": "string",
                    "example": "72a30ffb-1896-48b1-b006-985fb055db0f"
                }
            }
        },
...
        "Deck": {
            "type": "object",
            "properties": {
                "content": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Card"
                    }
                },
                "deckId": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "DeckResponse": {
            "type": "object",
            "properties": {
                "deck": {
                    "$ref": "#/definitions/Deck"
                },
                "message": {
                    "type": "string",
                    "example": "Successfully"
                }
            }
        },
...
        "HTTPError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer",
                    "example": 400
                },
                "message": {
                    "type": "string",
                    "example": "status bad request"
                }
            }
        },
...
        "SyncCardsRequest": {
            "type": "object",
            "required": [
                "cards"
            ],
            "properties": {
                "cards": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Card"
                    }
                }
            }
        },
...
        "UpdateDeckRequest": {
            "type": "object",
            "required": [
                "name"
            ],
            "properties": {
                "accessType": {
                    "type": "string",
                    "example": "public"
                },
                "name": {
                    "type": "string",
                    "example": "My new deck"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BasicAuth": {
            "type": "basic"
        },
        "Bearer": {
            "description": "ATTENTION! HOW TO USE: Type \"Bearer\" followed by a space and a token. Example: \"Bearer \\\u003ctoken\\\u003e\".",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    },
...
}